// src/content/listing-type.ts
export type Listing = {
  mls: string
  address: string
  city: string
  state: string
  zip: string
  price: string
  beds: string
  baths: string
  sqft?: string
  acres?: string
  yearBuilt?: string
  subdivision?: string
  county?: string
  propertyType?: string
  remarks?: string
  photo: string
  status?: string
  lat?: number
  lon?: number
}